Abstract publication feed into reusable component
Feed switching is handled on the page, and display is handled by the component.  The contents of the component are determined by the `relays` prop.
